requirements.txt
av
accelerate
config
cloudpickle
diffusers
einops
fastapi
fvcore
imageio
matplotlib
numpy
omegaconf
onnxruntime
opencv-python
pandas
pillow
pycocotools
pydantic
pyyaml
psutil
peft
python-multipart
regex==2024.5.15
safetensors
scikit-image
scipy
timm
tokenizers
torch
torchaudio
torchmetrics
torchvision
tqdm
transformers
uvicorn
runpod
asyncio